+++ llmexer/configs.py
@@ -2,8 +2,16 @@ import os
 
 from rich.console import Console
 
-from llmexer.common import GlobalFlags
 from llmexer.logger import get_logger
+
+
+class GlobalFlags:
+    """Class to hold global configuration state."""
+
+    dry_run: bool = False
+    verbose: bool = False
+    experiment_id: str = None
+
 
 logger = get_logger()
 console = Console()